Fixed dates cannot set before 1901-12-13 on 32bit
The problem with $date->getTimestamp() is that if the date is before 1901-12-13, it will return null on 32bit system. $datetimeplus will then be erroneously set to 1970-01-01.

This is evident during the unit test of Drupal\Tests\Component\Datetime\DateTimePlusTest::testValidateFormat() when it attempted to put in date as "11-03-31 17:44:00" and expects the return value from DateTimePlus to be "0011-03-31 17:44:00". On a 32-bit, it will return "1970-01-01 00:00:00" instead.

======Output of Unit Test======
1) Drupal\Tests\Component\Datetime\DateTimePlusTest::testValidateFormat
Failed asserting that two strings are equal.
--- Expected
+++ Actual
@@ @@
-'0011-03-31 17:44:00'
+'1970-01-01 00:00:00'
========================

The propose change will detect if the timestamp is beyond the scope of a 32bit integer and created the datetimeplus differently. Otherwise, it will create the datetimeplus like previously coded.